@@ -1,5 +1,6 @@
-import Vue from 'vue';
+import { defineComponent, onUnmounted, onMounted, ref, h } from 'vue';
 import {
+    getOptions,
     OptionTypes,
     create,
     supported,
@@ -42,14 +43,12 @@ const props = {};
 // Events that need to be mapped to emitters
 const events = [];
 
-// Props to watch
-const watch = {};
-
 // all active instances
 const instances = [];
 
 // global options
 let globalOptions = {};
+
 export const setOptions = (options) => {
     globalOptions = Object.assign(
         globalOptions,
@@ -61,12 +60,12 @@ export const setOptions = (options) => {
 };
 
 export default (...plugins) => {
-
     // register plugins in FilePond
     registerPlugin(...plugins);
 
     // build events and props array
     events.length = 0;
+
     for (const prop in OptionTypes) {
         // don't add events to the props array
         if (/^on/.test(prop)) {
@@ -75,135 +74,130 @@ export default (...plugins) => {
         }
 
         // get property type ( can be either a String or the type defined within FilePond )
-        props[prop] = [String, getNativeConstructorFromType(OptionTypes[prop])];
-
-        // setup watcher
-        watch[prop] = function (value) {
-            this._pond[prop] = value;
+        props[prop] = {
+            type: [String, getNativeConstructorFromType(OptionTypes[prop])],
+            default: '__undefined__',
         };
     }
 
     // create 
-    return Vue.component('FilePond', {
+    return defineComponent({
         name: 'FilePond',
         props,
-        watch,
-        render (h) {
-            return h('div',
+        setup(props, { emit, attrs }) {
+            const root = ref(null);
+            let pond = null;
+
+            // Will setup FilePond instance when mounted
+            onMounted(() => {
+                // exit here if not supported
+                if (!isSupported) return;
+
+                // get pond element
+                const element = root.value.querySelector('input');
+
+                // Map FilePond callback methods to Vue $emitters
+                const methods = events.reduce((obj, value) => {
+                    obj[value] = (...args) => {
+                        emit('input', pond ? pond.getFiles() : []);
+                        emit(value.substr(2), ...args);
+                    };
+                    return obj;
+                }, {});
+
+                const options = getOptions();
+                const userProps = {};
+
+                for (const prop in props) {
+                    if (props[prop] !== '__undefined__') {
+                        userProps[prop] = props[prop];
+                    }
+                }
+
+                const filePondOptions = Object.assign({},
+                    globalOptions,
+                    methods,
+                    attrs,
+                    options,
+                    userProps,
+                );
+
+                // Create our pond
+                pond = create(element, filePondOptions);
+
+                // Copy instance method references to component instance
+                Object.keys(pond)
+                    .filter(key => !filteredComponentMethods.includes(key))
+                    .forEach(key => root.value[key] = pond[key]);
+
+                // Add to instances so we can apply global options when used
+                instances.push(pond);
+
+                emit('init', pond);
+            });
+
+            // called when the component root node has been detached
+            const detached = () => {
+                // exit when no pond defined
+                if (!pond.value) return;
+
+                // bye bye pond
+                pond.value.destroy();
+
+                // remove from instances
+                const index = instances.indexOf(pond.value);
+                if (index >= 0) instances.splice(index, 1);
+
+                // clear reference
+                pond.value = null;
+            };
+
+            // Will clean up FilePond instance when unmounted
+            onUnmounted(() => {
+
+                // no longer attached, clean up
+                if (!root.value.offsetParent) {
+                    detached();
+                    return;
+                }
+
+                // if we're still attached it's likely a transition is running, we need to 
+                // determine the moment when we're no longer attached to the DOM so we can 
+                // clean up properly
+                const mutationHandler = (mutations, observer) => {
+                    const removedNodes = (mutations[0] || {}).removedNodes || [];
+                    const removedNode = removedNodes[0];
+                    if (!removedNode || !removedNode.contains(root.value)) return;
+                    observer.disconnect();
+                    detached();
+                }
+
+                // start observing parent element for changes to the DOM
+                const observer = new MutationObserver(mutationHandler);
+                observer.observe(document.documentElement, { childList: true, subtree: true });
+            });
+
+            return () => h('div',
                 {
+                    ref: root,
                     'class': {
                         'filepond--wrapper': true
                     }
                 },
                 [
-                    h('input', {
-                        attrs: {
-                            id: this.id,
-                            name: this.name,
-                            type: 'file',
-                            'class': this.className,
-                            required: this.required,
-                            multiple: this.allowMultiple,
-                            accept: this.acceptedFileTypes,
-                            capture: this.captureMethod
-                        }
+                    h('input',
+                    {
+                        id: props.id,
+                        name: props.name,
+                        type: 'file',
+                        'class': props.className,
+                        required: props.required,
+                        multiple: props.allowMultiple,
+                        accept: props.acceptedFileTypes,
+                        capture: props.captureMethod
                     })
                 ]
             );
-        },
-
-        // Will setup FilePond instance when mounted
-        mounted () {
-            // exit here if not supported
-            if (!isSupported) {
-                return;
-            }
-    
-            // get pond element
-            this._element = this.$el.querySelector('input');
-    
-            // Map FilePond callback methods to Vue $emitters
-            const options = events.reduce((obj, value) => {
-                obj[value] = (...args) => {
-                    this.$emit('input', this._pond ? this._pond.getFiles() : []);
-                    this.$emit(value.substr(2), ...args);
-                };
-                return obj;
-            }, {});
-    
-            // Scoop up attributes that might not have been caught by Vue ( because the props object is extended dynamically )
-            const attrs = Object.assign({}, this.$attrs);
-
-            // Create our pond
-            this._pond = create(
-                this._element,
-                Object.assign(
-                    {},
-                    globalOptions,
-                    options, 
-                    attrs, 
-                    this.$options.propsData
-                )
-            );
-            
-            // Copy instance method references to component instance
-            Object.keys(this._pond)
-                .filter(key => !filteredComponentMethods.includes(key))
-                .forEach(key => {
-                    this[key] = this._pond[key];
-                });
-
-            // Add to instances so we can apply global options when used
-            instances.push(this._pond);
-    
-        },
-    
-        // Will clean up FilePond instance when unmounted
-        destroyed () {
-
-            // reference to detached method
-            const { detached } = this.$options;
-            
-            // no longer attached, clean up
-            if (!this.$el.offsetParent) {
-                detached.call(this);
-                return;
-            }
-
-            // if we're still attached it's likely a transition is running, we need to 
-            // determine the moment when we're no longer attached to the DOM so we can 
-            // clean up properly
-            const mutationHandler = (mutations, observer) => {
-                const removedNodes = (mutations[0] || {}).removedNodes || [];
-                const removedNode = removedNodes[0];
-                if (!removedNode || !removedNode.contains(this.$el)) return;
-                observer.disconnect();
-                detached.call(this);
-            }
-
-            // start observing parent element for changes to the DOM
-            const observer = new MutationObserver(mutationHandler);
-            observer.observe(document.documentElement, { childList: true, subtree: true });
-        },
-
-        // called when the component root node has been detached
-        detached () {
-
-            // exit when no pond defined
-            if (!this._pond) return;
-    
-            // bye bye pond
-            this._pond.destroy();
-
-            // remove from instances
-            const index = instances.indexOf(this._pond);
-            if (index >= 0) {
-                instances.splice(index, 1);
-            }
-
-            // clear reference
-            this._pond = null;
         }
     });
-
 };
